Basic OllamaChatGPT/requirements.txt
langchain
langchain-core
langchain_community
python-dotenv
streamlit